修改违规处罚模块主键类型为Long

+++ src/main/java/com/gugugou/provider/violations/controller/ViolationsController.java
@@ -27,7 +27,7 @@ public class ViolationsController {
      * @return
      */
     @GetMapping("ticket/getTicketById")
-    public Violations getTicketById(@RequestParam Integer id){
+    public Violations getTicketById(@RequestParam Long id){
 
         logger.info("查询单条违规处罚单的入参：{}",id);
         return violationsService.getTicketById(id);
